requirements.txt
streamlit
pdfplumber
python-docx
scikit-learn
pandas
numpy
sentence-transformers